requirements.txt
amqp==5.3.1
asgiref==3.7.2
asttokens==3.0.0
async-timeout==5.0.1
bidict==0.23.1
billiard==4.2.1
boto3==1.28.15
botocore==1.31.15
CacheControl==0.14.0
cachetools==5.3.2
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.2
decorator==5.1.1
Django==3.2.12
django-cors-headers==4.3.1
django-crontab==0.7.1
django-elasticsearch-dsl==8.0
django-storages==1.14.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.5.0
elastic-transport==8.15.1
elasticsearch==8.16.0
elasticsearch-dsl==8.16.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
firebase-admin==6.3.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
ipython==8.18.1
jedi==0.19.2
jmespath==1.0.1
kombu==5.4.2
matplotlib-inline==0.1.7
msgpack==1.0.7
numpy==2.0.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
Pillow==10.0.0
prompt_toolkit==3.0.50
proto-plus==1.23.0
protobuf==4.25.2
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.19.1
PyJWT==2.8.0
pymongo==4.10.1
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2023.3.post1
razorpay==1.4.1
redis==5.2.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
rsa==4.9
s3transfer==0.6.2
simple-websocket==1.1.0
six==1.16.0
slack_sdk==3.34.0
sqlparse==0.2.4
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
wsproto==1.2.0
